+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - (New) Irace plots: generate an HTML report with the results of running the autoconfig procedure.
 - (New) Two ParetoSet implementations are provided: a simple list based (should be correct, but slow) and a NDTree based implementation (should be fast, but may contain bugs).
 - (New) ValidationResult: allow accumulating errors.
+- (New) PythonUtil: easily run Python commands inside a virtual environment
 - (New) When we report an unhandled exception, filter some stackframes that are not relevant to the user code. Full stacktrace still logged at trace level.
 - (Breaking) Due to changes in how objectives are handled, ReferenceResult methods have been renamed for clarity.
 - (Breaking) Removed Improver::_improve, please implement Improver::improve directly instead. To migrate, just rename the method and make it public.